Ignore empty files on import.
Prevent empty files to be imported, as they will lock the filestore and prevent project compilation.

While it might be better to try and fix the filestore issue, it is much easier to just prevent the issue by not importing empty files at all.

I don't see any kind of scenario where an empty file might be used in a latex project, so I don't feel it is a huge deal not to import them.