Replace event broadcasting with proper a scope.
Using a global event to update the object display overlay was a bad idea, so change that to use a proper scope on the directive and propagate the information through the scope.

The Controller will now $watch() for the value in scope and query the server for information about the object.